TST/REF: collect io tests by method
