+++ pandas/tests/io/test_pickle.py
@@ -23,13 +23,14 @@ import shutil
 from warnings import catch_warnings, simplefilter
 import zipfile
 
+import numpy as np
 import pytest
 
 from pandas.compat import PY38, get_lzma_file, import_lzma, is_platform_little_endian
 import pandas.util._test_decorators as td
 
 import pandas as pd
-from pandas import Index
+from pandas import Index, Series, period_range
 import pandas._testing as tm
 
 from pandas.tseries.offsets import Day, MonthEnd
@@ -548,3 +549,25 @@ def test_pickle_dataframe_with_multilevel_index(
     _test_roundtrip(frame.T)
     _test_roundtrip(ymd)
     _test_roundtrip(ymd.T)
+
+
+def test_pickle_timeseries_periodindex():
+    # GH#2891
+    prng = period_range("1/1/2011", "1/1/2012", freq="M")
+    ts = Series(np.random.randn(len(prng)), prng)
+    new_ts = tm.round_trip_pickle(ts)
+    assert new_ts.index.freq == "M"
+
+
+@pytest.mark.parametrize(
+    "name", [777, 777.0, "name", datetime.datetime(2001, 11, 11), (1, 2)]
+)
+def test_pickle_preserve_name(name):
+    def _pickle_roundtrip_name(obj):
+        with tm.ensure_clean() as path:
+            obj.to_pickle(path)
+            unpickled = pd.read_pickle(path)
+        return unpickled
+
+    unpickled = _pickle_roundtrip_name(tm.makeTimeSeries(name=name))
+    assert unpickled.name == name
